eancode modified

+++ app/src/main/java/com/example/betternutritions/model/AggregatedOrigin.kt
@@ -1,8 +1,8 @@
 package com.example.betternutritions.model
 
 data class AggregatedOrigin(
-    val epi_score: Int,
+    val epi_score: Double,
     val origin: String,
-    val percent: Int,
+    val percent: Double,
     val transportation_score: Any
 )
