+++ app/src/main/java/com/example/betternutritions/model/EcoscoreData.kt
@@ -4,7 +4,7 @@ data class EcoscoreData(
     val adjustments: Adjustments,
     val agribalyse: Agribalyse,
     val missing: Missing,
-    val missing_agribalyse_match_warning: Int,
+    val missing_agribalyse_match_warning: Double,
     val scores: Scores,
     val status: String
 )
