+++ app/src/main/java/com/example/betternutritions/model/Full.kt
@@ -1,6 +1,6 @@
 package com.example.betternutritions.model
 
 data class Full(
-    val h: Int,
-    val w: Int
+    val h: Double,
+    val w: Double
 )
