+++ app/src/main/java/com/example/betternutritions/model/IngredientX.kt
@@ -5,7 +5,7 @@ data class IngredientX(
     val ingredients: List<IngredientXX>,
     val percent_estimate: Double,
     val percent_max: Double,
-    val percent_min: Int,
+    val percent_min: Double,
     val text: String,
     val vegan: String,
     val vegetarian: String
