+++ app/src/main/java/com/example/betternutritions/model/IngredientXX.kt
@@ -4,7 +4,7 @@ data class IngredientXX(
     val id: String,
     val percent_estimate: Double,
     val percent_max: Double,
-    val percent_min: Int,
+    val percent_min: Double,
     val text: String,
     val vegan: String,
     val vegetarian: String
