+++ app/src/main/java/com/example/betternutritions/model/Ingredients.kt
@@ -1,7 +1,7 @@
 package com.example.betternutritions.model
 
 data class Ingredients(
-    val dummy: Int
+    val dummy: Double
     /*
     val display: DisplayX,
     val small: SmallX,
