+++ app/src/main/java/com/example/betternutritions/model/IngredientsAnalysis.kt
@@ -1,7 +1,7 @@
 package com.example.betternutritions.model
 
 data class IngredientsAnalysis(
-    val dummy: Int
+    val dummy: Double
     /*val en:may-contain-palm-oil: List<String>,
     val en:non-vegan: List<String>,
     val en:vegan-status-unknown: List<String>,
