+++ app/src/main/java/com/example/betternutritions/model/LanguagesCodes.kt
@@ -1,5 +1,5 @@
 package com.example.betternutritions.model
 
 data class LanguagesCodes(
-    val de: Int
+    val de: Double
 )
