+++ app/src/main/java/com/example/betternutritions/model/Missing.kt
@@ -1,6 +1,6 @@
 package com.example.betternutritions.model
 
 data class Missing(
-    val agb_category: Int,
-    val origins: Int
+    val agb_category: Double,
+    val origins: Double
 )
