+++ app/src/main/java/com/example/betternutritions/model/Packaging.kt
@@ -1,8 +1,8 @@
 package com.example.betternutritions.model
 
 data class Packaging(
-    val non_recyclable_and_non_biodegradable_materials: Int,
+    val non_recyclable_and_non_biodegradable_materials: Double,
     val packagings: List<PackagingX>,
-    val score: Int,
-    val value: Int
+    val score: Double,
+    val value: Double
 )
