+++ app/src/main/java/com/example/betternutritions/model/PackagingX.kt
@@ -1,7 +1,7 @@
 package com.example.betternutritions.model
 
 data class PackagingX(
-    val ecoscore_material_score: Int,
+    val ecoscore_material_score: Double,
     val ecoscore_shape_ratio: Double,
     val material: String,
     val non_recyclable_and_non_biodegradable: String,
