+++ app/src/main/java/com/example/betternutritions/model/SizesX.kt
@@ -1,7 +1,7 @@
 package com.example.betternutritions.model
 
 data class SizesX(
-    val dummy: Int
+    val dummy: Double
     /*val `100`: X100,
     val `400`: X100,
     val full: FullX
