+++ app/src/main/java/com/example/betternutritions/model/SizesXXX.kt
@@ -1,7 +1,7 @@
 package com.example.betternutritions.model
 
 data class SizesXXX(
-    val dummy: Int
+    val dummy: Double
     /*
     val `100`: X100,
     val `200`: X100,
