+++ app/src/main/java/com/example/betternutritions/model/X1.kt
@@ -2,6 +2,6 @@ package com.example.betternutritions.model
 
 data class X1(
     val sizes: Sizes,
-    val uploaded_t: Int,
+    val uploaded_t: Double,
     val uploader: String
 )
