+++ app/src/main/java/com/example/betternutritions/model/X2.kt
@@ -2,6 +2,6 @@ package com.example.betternutritions.model
 
 data class X2(
     val sizes: SizesX,
-    val uploaded_t: Int,
+    val uploaded_t: Double,
     val uploader: String
 )
