+++ app/src/main/java/com/example/betternutritions/model/X3.kt
@@ -2,6 +2,6 @@ package com.example.betternutritions.model
 
 data class X3(
     val sizes: SizesXX,
-    val uploaded_t: Int,
+    val uploaded_t: Double,
     val uploader: String
 )
